Rough version of create-hikey960-rootfs
* Creates a rootfs with debootstrap, sets hostname to hikey960,
  sets up /etc/fstab to load the correct partition, etc.
* Is still missing copying the kernel/modules, and doesn't create
  grub.cfg yet.

Once this is used to create a root filesystem, a sparse image file
(which should be no larger than 2G) should be created, an ext4 filesystem
created on that with mkfs.ext4, and the result converted to an simg
with ext2simg. This can then be flashed on the hikey960.